Support methods with macros in their bodies
