+++ TypedSyntax/test/runtests.jl
@@ -4,79 +4,7 @@ using InteractiveUtils, Test
 
 has_name_typ(node, name::Symbol, @nospecialize(T)) = kind(node) == K"Identifier" && node.val === name && node.typ === T
 
-module TSN
-
-# with two uses of the same slot in the same call. Must start on line 10 (or update test below)
-function simplef(a, b)
-    z = a * a
-    return z + b
-end
-
-function has2xa(x)
-    x &= x
-end
-function has2xb(x)
-    x -= x
-    return x
-end
-
-# This is taken from the definition of `sin(::Int)` in Base, copied here for testing purposes
-# in case the implementation changes
-for f in (:mysin,)
-    @eval function ($f)(x::Real)
-        xf = float(x)
-        x === xf && throw(MethodError($f, (x,)))
-        return ($f)(xf)
-    end
-end
-mysin(x::AbstractFloat) = sin(x)
-
-function summer(list)
-    s = 0                    # deliberately ::Int to test type-changes
-    for x in list
-        s += x
-    end
-    return s
-end
-
-zerowhere(::AbstractArray{T}) where T<:Real = zero(T)
-cb(a, i) = checkbounds(Bool, a, i)
-
-add2(x) = x[1] + x[2]
-
-myabs(x) = x < 0 ? -x : x
-
-likevect(X::T...) where {T} = T[ X[i] for i = 1:length(X) ]
-cbva(a, i...) = checkbounds(Bool, a, i...)
-anykwargs(; kwargs...) = println(kwargs...)
-splats(x, y) = vcat(x..., y...)
-
-myoftype(ref, val) = typeof(ref)(val)
-
-defaultarg(x, y=2) = x + y
-hasdefaulttypearg(::Type{T}=Rational{Int}) where T = zero(T)
-
-charset1 = 'a':'z'
-getchar1(idx) = charset1[idx]
-const charset2 = 'a':'z'
-getchar2(idx) = charset2[idx]
-
-# unused statements
-function mycheckbounds(A, i)
-    checkbounds(Bool, A, i) || Base.throw_boundserror(A, i)
-    return nothing
-end
-
-# Implementation of a struct & interface
-struct DefaultArray{T,N,A<:AbstractArray{T,N}} <: AbstractArray{T,N}
-    parentarray::A
-    defaultvalue::T
-end
-DefaultArray(parentarray, defaultvalue) = DefaultArray{ndims(parentarray)}(parentarray, defaultvalue)
-Base.getindex(a::DefaultArray{T,N}, i::Vararg{Int,N}) where {T,N} = checkbounds(Bool, a, i...) ? a.parentarray[i...] : a.defaultvalue
-Base.size(a::DefaultArray) = size(a.parentarray)
-
-end
+include("test_module.jl")
 
 @testset "TypedSyntax.jl" begin
     st = """
@@ -166,6 +94,12 @@ end
         @test child(body, idxsinner...).typ === nothing
         @test child(body, idxsouter...).typ === Vector{Real}
     end
+
+    # body macros
+    tsn = TypedSyntaxNode(TSN.hasmacro, (Tuple{Int,Int}, Char))
+    sig, body = children(tsn)
+    @test has_name_typ(child(body, 2, 2, 2), :t, Tuple{Int,Int})
+    @test has_name_typ(child(body, 2, 3), :x, Char)
 
     # `ref` indexing
     st = """
@@ -420,10 +354,10 @@ end
         printstyled(io, obj; hide_type_stable=false)
     end
     @test str === """
-        10 function simplef(a::Int64, b::Float64)::Float64
-        11     z::Int64 = (a::Int64 * a::Int64)::Int64
-        12     return (z::Int64 + b::Float64)::Float64
-        13 end"""
+        4 function simplef(a::Int64, b::Float64)::Float64
+        5     z::Int64 = (a::Int64 * a::Int64)::Int64
+        6     return (z::Int64 + b::Float64)::Float64
+        7 end"""
     tsn = TypedSyntaxNode(TSN.myabs, (Float64,))
     str = sprint(tsn; context=:color=>false) do io, obj
         printstyled(io, obj; hide_type_stable=false)
